refactor: final cleanup
