+++ .github/scripts/helpers/get-files.js
