+++ .github/scripts/locales-format-checker/check-locales.js
@@ -1,7 +1,14 @@
 import * as core from "@actions/core";
-import { COLORS, fileNameFormat, i18nextKeyExtensions, keyFormat, LOCALES_DIR, mainLanguage } from "./constants.js";
-import { getFiles, getKeys, getMainLanguageKeys, removeLanguageCode } from "./get-files.js";
-import { getCorrectFormat, printVerboseLog } from "./utils.js";
+import {
+  toCamelCase,
+  toKebabCase,
+  toPascalCase,
+  toPascalSnakeCase,
+  toSnakeCase,
+  toUpperSnakeCase,
+} from "../helpers/strings.js";
+import { COLORS, fileNameFormat, i18nextKeyExtensions, keyFormat, LOCALES_DIR, mainLanguage } from "../helpers/constants.js";
+import { getFiles, getKeys, getMainLanguageKeys, removeLanguageCode } from "../helpers/get-files.js";
 
 /** @import * from "./types" */
 
@@ -261,3 +268,39 @@ function checkForMissingKeys(filePath, options) {
 }
 
 //#endregion
+
+/**
+ * Returns the correct format for the provided format.
+ * @param {string} key - The key to get the correct format for.
+ * @param {format} format - The format to get the correct format for.
+ * @returns {string} The correct format.
+ */
+function getCorrectFormat(key, format) {
+  switch (format) {
+    case "camelCase":
+      return toCamelCase(key);
+    case "kebab-case":
+      return toKebabCase(key);
+    case "PascalCase":
+      return toPascalCase(key);
+    case "snake_case":
+      return toSnakeCase(key);
+    case "UPPER_SNAKE_CASE":
+      return toUpperSnakeCase(key);
+    case "Pascal_Snake_Case":
+      return toPascalSnakeCase(key);
+    default:
+      core.setFailed(`Unknown format: "${format}"`);
+  }
+}
+
+/**
+ * Prints a console log if verbose logging is enabled
+ * @param {string} text - The text to print to the console
+ * @param {options} options - The command line options.
+ */
+function printVerboseLog(text, options) {
+  if (options.verbose) {
+    core.info(text);
+  }
+}
